\name{ASexpression-package}
\alias{ASexpression-package}
\alias{ASexpression}
\docType{package}
\title{
R package for Allele-specific differential expression analysis
~~ ASexpression ~~
}
\description{
Takes as input the featureCounts output as a part of allele-specific mapping pipeline, and does the DE.
~~ R package for Allele-specific differential expression analysis using featureCounts output ~~
}
\details{
\tabular{ll}{
Package: \tab ASexpression\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-07\cr
License: \tab GPL 2.0\cr
}

}
\author{
Vivek Bhardwaj

Maintainer: Vivek Bhardwaj <bhardwaj@ie-freiburg.mpg.de>

}
\references{

}
\keyword{ package }
\seealso{
}
\examples{

}
